requirements.txt
arrow==0.13.0
autopep8==1.4.3
awscli==1.16.87
backcall==0.1.0
bleach==3.1.0
bokeh==1.0.4
botocore==1.12.77
certifi==2018.11.29
chardet==3.0.4
chartify==2.3.5
colorama==0.3.9
colour==0.1.5
comet-git-pure==0.19.9
configobj==5.0.6
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
docutils==0.14
elpy==1.28.0
entrypoints==0.3
epc==0.0.5
-e git+git@gitlab.com:jayemar/vsb_power_line.git@0b20a8f183dc3d7a9695917d1a75e3dc3de20039#egg=etl&subdirectory=etl
everett==1.0.1
flake8==3.7.2
idna==2.8
importmagic==0.1.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-full-width==1.2.0
jupyterthemes==0.20.0
kaggle==1.5.2
kiwisolver==1.0.1
lesscpy==0.13.0
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.9
notebook==5.7.4
numpy==1.16.0
nvidia-ml-py3==7.352.0
packaging==18.0
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
ply==3.11
progressbar2==3.39.2
prometheus-client==0.5.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pyarrow==0.12.0
pyasn1==0.4.5
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
python-slugify==2.0.1
python-utils==2.3.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
selenium==3.8.0
Send2Trash==1.5.0
sexpdata==0.0.3
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.29.0
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
widgetsnbextension==3.4.2
wurlitzer==1.0.2
